+++ backend/src/main/java/com/proghuly/cardapio/DTO/FoodResponseDTO.java
@@ -2,7 +2,7 @@ package com.proghuly.cardapio.DTO;
 
 import com.proghuly.cardapio.entity.Food;
 
-public record FoodResponseDTO(Long id, String title, String image, Integer price) {
+public record FoodResponseDTO(Long id, String title, String image, double price) {
 
     public FoodResponseDTO(Food food) {
         this(food.getId(), food.getTitle(), food.getImage(), food.getPrice());
